Extract synth orchestration from `HandleComponents` into a `HandleSynths` mechanism
